Update Program.cs
I am getting the base of the programm from https://www.geeksforgeeks.org/c/snake-game-in-c/.
With the base down I will split up the methods to classes so that there is a clear structure and is more readable and sustainable. From now on I will try to implement OOP (Object Oriented Programming) in this project.