Updated Body movement
In Program.cs I changed the iteration of snakeTailX and -Y in Logic() as it is easier for me to undestand and makes the traversal through the arraye better for the project. In Snake.cs I added a previousHeadPosition so that the body can follow the head correctly.

Problem now is that the body is still in some cases not following the body correctly.