Added Interface
I added an Interface IRendable, mainly so that Programm doesn't have acces to Snake and Fruit as it simply not needed. But because of this change the snake and fruit are flickering. That is because the space between the '#' borders when drawing the field is empty. So every frame when everything was drawned the space between '#' gets emptied.